// src/types.ts
export type NFTMetaType = {
    name: string
    tradeDirection: string
    price: number
    priceUSD: number
    image: string
    transactionDate: number
    marketPlaceURL: string
}

export type dataNFT = {
    ratio : string
    volume :number
    avergae: number
    listed: number
    floor: number
}